pyproject.toml
[tool.poetry]
name = "lg"
version = "1.0.0"
description = "Python Network Looking Glass."
authors = ["Rob Woodward <rob@emailplus.org>"]
license = "BSD-2-Clause"
readme = "README.md"
repository = "https://github.com/robwdwd/lg"
include = ["README.md", "LICENSE", "app.py", "examples"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = {extras = ["standard"], version = "^0.21.0"}
gunicorn = "^20.1.0"
starlette = "^0.26.0"
Jinja2 = "^3.1.2"
Starlette-WTF = "^0.4.3"
pyyaml = "^6.0"
ttp = "^0.9.2"
scrapli = {extras = ["asyncssh", "ttp"], version = "^2023.1.30"}

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.1.3"
pylint = "^2.9.6"
pydocstyle = "^6.1.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
